django_project/requirements.txt
django
djangorestframework
django_extensions
django-crispy-forms
django-cors-headers
psycopg2
xhtml2pdf
django-glrm
black